fix(load-video): decode every frame instead of a 64-frame preview
LoadVideoPathNode borrowed SaveVideoNode._extract_frames without a mode and
inherited its "preview (64)" default, so any clip over 64 frames was evenly
resampled and played back near double speed downstream. Decoding now goes
through core.media_converter.decode_video_frames() for both nodes; the
loader reports frame_count from the decoded tensor, and the save node's
_extract_frames requires an explicit mode. Save Video's images output
defaults to "all", and fps is a FLOAT so NTSC rates are expressible.

+++ core/video_compare.py
@@ -100,7 +100,7 @@ def combine_videos(
     labels: list[str] | None = None,
     gap: int = 4,
     bg_color: str = _DEFAULT_BG,
-    fps: int = 24,
+    fps: float = 24.0,
 ) -> str | None:
     """Stack ``sources`` into one comparison mp4 and return its path.
